libs/python-sdk/requirements.txt
rel==0.4.8
websocket-client==1.5.1
